Character tweaks
Normally when changing the data type of a stat variable and saving the script data of ALL resources would be lost. With this update, if the variable name is the same but you change the type, it'll restore the data of all OPENED character resources. So switching between RangeFloat and RangeInt will be harmless on opened characters. You'll be prompted to save opened sheets though.