Trying to fix a very elusive crash with the Tree node
Sometimes, when repopulating items in a Tree node when you hovered on an item the editor would crash. I believe this is an engine issue that has to do with how it clears TreeItem objects from memory when using TreeItem.free(); In an attempt at fixing it I'll switch to using Tree.clear() which seems to be slightly more reliable.